added functionality to viewFullSite link to open a link resource in new tab

+++ grails-app/domain/com/ig/LinkShare/ResourceRating.groovy
@@ -1,7 +1,7 @@
 package com.ig.LinkShare
 
 class ResourceRating {
-    Integer score
+    float score
 
     static belongsTo = [user:User,resource:Resource]
     static constraints = {
